+++ build.gradle
@@ -44,7 +44,7 @@ test {
 }
 
 application {
-    mainClassName = "Launcher"
+    mainClassName = "duke.gui.Launcher"
 }
 
 shadowJar {
