+++ data/taskList.txt
@@ -2,9 +2,11 @@
 2. | T | 1 | 1231
 3. | D | 1 | asdasd | 2021-08-13
 4. | T | 1 | wah
-5. | T | 0 | sdf
-6. | T | 1 | whw
-7. | T | 0 | 5
-8. | E | 1 | wah | 2021-08-11
-9. | E | 0 | zzz | 2021-08-11 | 13:12
-10. | E | 0 | zzz | 2021-08-11 | 11:12
+5. | T | 1 | whw
+6. | T | 0 | 5
+7. | E | 1 | wah | 2021-08-11
+8. | E | 0 | zzz | 2021-08-11 | 13:12
+9. | E | 0 | zzz | 2021-08-11 | 11:12
+10. | T | 0 | afdasd
+11. | T | 0 | damn
+12. | T | 0 | sliceofmallow
