+++ src/main/java/duke/gui/Launcher.java
@@ -1,3 +1,5 @@
+package duke.gui;
+
 import javafx.application.Application;
 
 /**
