+++ src/main/java/duke/gui/Main.java
@@ -1,3 +1,5 @@
+package duke.gui;
+
 import java.io.IOException;
 
 import javafx.application.Application;
@@ -22,6 +24,7 @@ public class Main extends Application {
             AnchorPane ap = fxmlLoader.load();
             Scene scene = new Scene(ap);
             stage.setScene(scene);
+            stage.setTitle("Herbert!");
             fxmlLoader.<MainWindow>getController().setDuke(duke);
             stage.show();
         } catch (IOException e) {
