+++ src/main/java/duke/gui/MainWindow.java
@@ -1,3 +1,5 @@
+package duke.gui;
+
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.TimeUnit;
