+++ src/main/resources/view/DialogBox.fxml
@@ -7,7 +7,12 @@
 
 <fx:root alignment="TOP_RIGHT" maxHeight="1.7976931348623157E308" maxWidth="1.7976931348623157E308" prefWidth="400.0" type="javafx.scene.layout.HBox" xmlns="http://javafx.com/javafx/8.0.171" xmlns:fx="http://javafx.com/fxml/1">
     <children>
-        <Label fx:id="dialog" text="Label" wrapText="true" />
+        <!-- Style attribute adapted from https://github.com/royleochan/ip -->
+        <Label fx:id="dialog" style="-fx-background-color: #b794f6; -fx-label-padding: 7; -fx-background-radius: 12; -fx-padding: 5;" text="Label" wrapText="true" minHeight="-Infinity">
+            <HBox.margin>
+                <Insets left="18.0" />
+            </HBox.margin>
+        </Label>
         <ImageView fx:id="displayPicture" fitHeight="99.0" fitWidth="99.0" pickOnBounds="true" preserveRatio="true" />
     </children>
     <padding>
